move commune filler command to fixtures
